// src/components/type.ts
import {FlexAlignType} from 'react-native';

export type InjectProps = {
  m?: number;
  mt?: number;
  mr?: number;
  mb?: number;
  ml?: number;
  mv?: number;
  mh?: number;
  p?: number;
  pt?: number;
  pr?: number;
  pb?: number;
  pl?: number;
  pv?: number;
  ph?: number;
  alignSelf?: FlexAlignType;
  zIndex?: number;
  flex?: number;
  width?: number | string;
  height?: number | string;
  backgroundColor?: string;
};